+++ WebApiStructureNetCore/Models/CustomExceptionResponse.cs
@@ -2,7 +2,7 @@
 {
     public class CustomExceptionResponse
     {
-        public int TipoErro { get; set; }
+        public short TipoErro { get; set; }
         public string Mensagem { get; set; }
         public string Id { get; set; }
     }
